requirements.txt
psycopg2-binary==2.9.9
fastapi==0.110.0
uvicorn==0.27.1 
python-dotenv==1.0.1 
pandas==2.2.1
tqdm==4.66.2 